+++ docs/docs/release_notes.md
@@ -1,6 +1,7 @@
 ### 3.2.4
 - Add all() query
 - Add tests for `pantherdb`, `load_configs()`, `status.py`, `Panel`, `multipart`, `request headers`
+- Refactor `Headers()` class
 - Minor Improvement
 
 ### 3.2.1